WIP-HOLD Navigate menu via highlight chars
Add char highlight for key-sequence commands

For example _F_ile would highlight the F, possibly
by underscoring it, and if the user pressed the sequence
F10-f
then the menu item would activate (assuming F10 activated the menu)

Similar f10-f-o, would go first to _F_ile, then to _O_pen

This makes it easy for users to learn long sequence of keys
for a command, because they can use their spartial memory.

It is important for spartial memory that all menu items are 
always located the same place. If the menu should react to
context, then it is better to disable a menu item 
(rendered as grey), than to hide it (not rendered)